api/requirements.txt
fastapi
pinecone-client
torch
numpy
pandas
scipy
implicit
tqdm
joblib
kaggle